build: Rework posix dependencies
Centralize where are they processed, both for the smoke tests and for
unit tests, and remove therefore the mocks (which where no mocks but the
actual dependencies with CMake wrapping)

For cJSON we require the system package while for Mbed TLS we fetch it
from upstream. See comment.

Note that for Mbed TLS we were already using the fetched one (via the
mock), so we have never been using the system package. Modified the
README to reflect this.

Ticket: MEN-7526

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>